requirements.txt
astroid==2.3.3
Babel==2.8.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
bootstrap4==0.1.0
cffi==1.13.2
colorama==0.4.3
Django==2.2
django-bootstrap==0.2.4
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap3==12.0.3
django-bootstrap4==1.1.1
django-contrib-comments==1.9.2
django-countries==5.5
django-crispy-forms==1.8.1
django-datetimepicker==3.14
django-debug-toolbar==2.2
django-phonenumber-field==4.0.0
Faker==4.0.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==1.4.6
phonenumbers==8.11.2
Pillow==7.0.0
pycparser==2.19
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
reusables==0.9.5
termcolor==1.1.0
text-unidecode==1.3
typed-ast==1.4.1
Unidecode==1.1.1
wrapt==1.11.2
